pyproject.toml
[tool.poetry]
name = "trader-bot"
version = "0.1.0"
description = "Trading bot for CB Pro"
authors = ["diznq <13458283+diznq@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
cbpro = "^1.1.4"
PyYAML = "^6.0"
redis = "^4.1.0"
pandas = "^1.3.5"
pydantic = "^1.9.0"
urllib3 = "^1.26.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
